Moved graphql_mutation to graphql_core.
