+++ modules/graphql_core/src/Plugin/GraphQL/Fields/Mutations/ConstraintViolationMessage.php
@@ -1,31 +1,31 @@
 <?php
 
-namespace Drupal\graphql_mutation\Plugin\GraphQL\Fields;
+namespace Drupal\graphql_core\Plugin\GraphQL\Fields\Mutations;
 
 use Drupal\graphql\Plugin\GraphQL\Fields\FieldPluginBase;
 use Symfony\Component\Validator\ConstraintViolationInterface;
 use Youshido\GraphQL\Execution\ResolveInfo;
 
 /**
- * Retrieve the error code of an entity constraint violation.
+ * Retrieve the message of an entity constraint violation.
  *
  * @GraphQLField(
- *   id = "constraint_violation_code",
+ *   id = "constraint_violation_message",
  *   secure = true,
- *   name = "code",
+ *   name = "message",
  *   type = "String",
  *   types = {"ConstraintViolation"},
  *   nullable = true
  * )
  */
-class ConstraintViolationCode extends FieldPluginBase {
+class ConstraintViolationMessage extends FieldPluginBase {
 
   /**
    * {@inheritdoc}
    */
   public function resolveValues($value, array $args, ResolveInfo $info) {
-    if ($value instanceof ConstraintViolationInterface) {
-      yield $value->getCode();
+    if ($value instanceof ConstraintViolationInterface){
+      yield $value->getMessage();
     }
   }
