+++ modules/graphql_mutation/graphql_mutation.info.yml
@@ -3,5 +3,6 @@ type: module
 description: 'Core module with common fields and types for enabling mutations.'
 package: GraphQL
 core: 8.x
+hidden: true
 dependencies:
   - graphql_core
